+++ mrv2/lib/mrvPy/Cmds.h
@@ -19,6 +19,7 @@ extern void mrv2_settings(py::module& m);
 extern void mrv2_filesmodel(py::module& m);
 extern void mrv2_playlist(py::module& m);
 extern void mrv2_io(py::module& m);
+extern void mrv2_annotations(py::module& m);
 extern void mrv2_commands(py::module& m);
 extern void mrv2_python_plugins(py::module& m);
 extern void mrv2_discover_python_plugins();